+++ controllers/role_utils.py
@@ -1,5 +1,5 @@
 from functools import wraps
-from dbcontext.dbcontext import get_user_by_id
+from dbcontext.userdata import get_user_by_id
 
 # Define role hierarchy
 ROLE_HIERARCHY = {
